Added some test dependencies, renamed socketEvents.js
