+++ package.json
@@ -7,7 +7,7 @@
     "start": "node src/app.js",
     "lint": "eslint --ignore-path .gitignore .",
     "fix-lint": "eslint --fix --ignore-path .gitignore .",
-    "test": "mocha ./test/* --exit"
+    "test": "mocha ./tests/* --exit"
   },
   "repository": {
     "type": "git",
@@ -30,7 +30,10 @@
     "mquery": "^3.2.3"
   },
   "devDependencies": {
+    "chai": "^4.2.0",
     "eslint": "^7.8.1",
-    "mocha": "^8.2.1"
+    "mocha": "^8.2.1",
+    "proxyrequire": "^1.0.21",
+    "sinon": "^9.2.2"
   }
 }
