+++ src/app.js
@@ -1,7 +1,7 @@
 const zmq = require('zeromq');
 const sock = zmq.socket('sub');
 const schedule = require('node-schedule');
-const socketEvents = require('./events/socketEvents');
+const socketEvents = require('./events/socket-events');
 require('console-stamp')(console, '[HH:MM:ss.l]');
 
 const address = 'tcp://eddn.edcd.io:9500';
